+++ compiler-cli/src/lsp.rs
@@ -15,10 +15,10 @@ use gleam_core::{
     build::{self, Located, Module, ProjectCompiler},
     config::PackageConfig,
     diagnostic::{self, Level},
-    io::{CommandExecutor, FileSystemIO},
+    io::{CommandExecutor, FileSystemIO, Stdio},
     line_numbers::LineNumbers,
     paths,
-    type_::{pretty::Printer, HasType},
+    type_::pretty::Printer,
     Error, Result,
 };
 use itertools::Itertools;
@@ -744,7 +744,7 @@ impl LanguageServer {
 
             // Otherwise format the file from disc
             None => {
-                let src = crate::fs::read(&path)?;
+                let src = crate::fs::read(path)?;
                 gleam_core::format::pretty(&mut new_text, &src, Path::new(path))?;
             }
         };
@@ -795,7 +795,7 @@ fn uri_to_module_name_test() {
 fn uri_to_module_name(uri: &Url, root: &Path) -> Option<String> {
     let path = PathBuf::from(uri.path());
     let components = path
-        .strip_prefix(&root)
+        .strip_prefix(root)
         .ok()?
         .components()
         .skip(1)
@@ -1011,7 +1011,7 @@ where
             ProjectCompiler::new(config, options, manifest.packages, Box::new(telemetry), io);
         // To avoid the Erlang compiler printing to stdout (and thus
         // violating LSP which is currently using stdout) we silence it.
-        project_compiler.silence_subprocess_stdout = true;
+        project_compiler.subprocess_stdio = Stdio::Null;
 
         Ok(Self {
             project_compiler,