+++ compiler-cli/src/run.rs
@@ -2,7 +2,7 @@ use gleam_core::{
     build::{Mode, Options, Target},
     config::PackageConfig,
     error::Error,
-    io::CommandExecutor,
+    io::{CommandExecutor, Stdio},
     paths,
 };
 
@@ -67,7 +67,7 @@ fn run_erlang(module: &str, arguments: Vec<String>) -> Result<i32, Error> {
         args.push(argument);
     }
 
-    ProjectIO::new().exec("erl", &args, &[], None, false)
+    ProjectIO::new().exec("erl", &args, &[], None, Stdio::Inherit)
 }
 
 fn run_javascript(
@@ -94,5 +94,5 @@ fn run_javascript(
         args.push(argument);
     }
 
-    ProjectIO::new().exec("node", &args, &[], None, false)
+    ProjectIO::new().exec("node", &args, &[], None, Stdio::Inherit)
 }