+++ compiler-core/src/format/tests.rs
@@ -3706,3 +3706,133 @@ pub external type Map(key, value)
 "#
     );
 }
+
+#[test]
+fn use_1() {
+    assert_format!(
+        r#"pub fn main() {
+  use <- benchmark("thingy")
+}
+"#
+    );
+}
+
+#[test]
+fn use_2() {
+    assert_format!(
+        r#"pub fn main() {
+  use user <- login()
+}
+"#
+    );
+}
+
+#[test]
+fn use_3() {
+    assert_format!(
+        r#"pub fn main() {
+  use one, two, three, four <- get_multiple_things()
+}
+"#
+    );
+}
+
+#[test]
+fn use_4() {
+    assert_format!(
+        r#"pub fn main() {
+  use
+    one,
+    two,
+    three,
+    four,
+    five,
+    six,
+    seven,
+    eight,
+    nine,
+    ten,
+    eleven
+  <- get_multiple_things_with_a_longer_function
+}
+"#
+    );
+}
+
+#[test]
+fn use_5() {
+    assert_format!(
+        r#"pub fn main() {
+  use
+    one,
+    two,
+    three,
+    four,
+    five,
+    six,
+    seven,
+    eight,
+    nine,
+    ten,
+    eleven
+  <- get_multiple_things_with_a_longer_function(a, b, c, d)
+}
+"#
+    );
+}
+
+#[test]
+fn use_6() {
+    assert_format!(
+        r#"pub fn main() {
+  use
+    one,
+    two,
+    three,
+    four,
+    five,
+    six,
+    seven,
+    eight,
+    nine,
+    ten,
+    eleven
+  <- get_multiple_things_with_a_longer_function(
+      "one",
+      "two",
+      "three",
+      "four",
+      "five",
+      "six",
+      "seven",
+      "eight",
+    )
+}
+"#
+    );
+}
+
+#[test]
+fn use_pipe_call() {
+    assert_format!(
+        r#"pub fn main() {
+  use <- a
+    |> b
+  c
+}
+"#
+    );
+}
+#[test]
+fn use_pipe_everything() {
+    assert_format!(
+        r#"pub fn main() {
+  {
+    use <- a
+  }
+  |> b
+  c
+}
+"#
+    );
+}