+++ compiler-core/src/parse.rs
@@ -60,7 +60,7 @@ use crate::ast::{
     RecordConstructor, RecordConstructorArg, RecordUpdateSpread, SrcSpan, Statement, TargetGroup,
     TodoKind, TypeAst, UnqualifiedImport, UntypedArg, UntypedClause, UntypedClauseGuard,
     UntypedConstant, UntypedExpr, UntypedExternalFnArg, UntypedModule, UntypedPattern,
-    UntypedRecordUpdateArg, UntypedStatement, CAPTURE_VARIABLE,
+    UntypedRecordUpdateArg, UntypedStatement, Use, CAPTURE_VARIABLE,
 };
 use crate::build::Target;
 use crate::parse::extra::ModuleExtra;
@@ -537,6 +537,11 @@ where
                 self.parse_assignment(start, AssignmentKind::Assert)?
             }
 
+            Some((start, Token::Use, _)) => {
+                let _ = self.next_tok();
+                self.parse_use(start)?
+            }
+
             // helpful error on possibly trying to group with ""
             Some((start, Token::LeftParen, _)) => {
                 return parse_error(ParseErrorType::ExprLparStart, SrcSpan { start, end: start });
@@ -676,6 +681,35 @@ where
         }
 
         Ok(Some(expr))
+    }
+
+    // A `use` expression
+    // use <- function
+    // use <- function()
+    // use <- function(a, b)
+    // use <- module.function(a, b)
+    // use a, b, c <- function(a, b)
+    // use a, b, c, <- function(a, b)
+    fn parse_use(&mut self, start: u32) -> Result<UntypedExpr, ParseError> {
+        let assignments = if let Some((_, Token::LArrow, _)) = self.tok0 {
+            vec![]
+        } else {
+            Parser::series_of(self, &Parser::parse_use_assignment, Some(&Token::Comma))?
+        };
+
+        _ = self.expect_one(&Token::LArrow)?;
+        let call = self.expect_expression()?;
+
+        Ok(UntypedExpr::Use(Use {
+            location: SrcSpan::new(start, call.location().end),
+            assignments,
+            call: Box::new(call),
+        }))
+    }
+
+    fn parse_use_assignment(&mut self) -> Result<Option<(AssignName, SrcSpan)>, ParseError> {
+        let (start, name, end) = self.expect_assign_name()?;
+        Ok(Some((name, SrcSpan::new(start, end))))
     }
 
     // An assignment, with `Let` or `Assert` already consumed
@@ -992,7 +1026,7 @@ where
         if let Some(lead) = &patterns.first() {
             let mut alternative_patterns = vec![];
             loop {
-                if None == self.maybe_one(&Token::Vbar) {
+                if self.maybe_one(&Token::Vbar).is_none() {
                     break;
                 }
                 alternative_patterns.push(self.parse_patterns()?);
@@ -1221,6 +1255,7 @@ where
                 let _ = self.next_tok();
                 if let Some(value) = self.parse_pattern()? {
                     Ok(Some(CallArg {
+                        implicit: false,
                         location: SrcSpan {
                             start,
                             end: value.location().end,
@@ -1244,6 +1279,7 @@ where
                 self.tok1 = t1;
                 if let Some(value) = self.parse_pattern()? {
                     Ok(Some(CallArg {
+                        implicit: false,
                         location: value.location(),
                         label: None,
                         value,
@@ -1548,6 +1584,7 @@ where
                 location.start = start
             };
             Ok(Some(ParserArg::Arg(Box::new(CallArg {
+                implicit: false,
                 label,
                 location,
                 value,
@@ -2231,6 +2268,7 @@ where
         if let Some(value) = self.parse_const_value()? {
             if let Some((start, label, _)) = name {
                 Ok(Some(CallArg {
+                    implicit: false,
                     location: SrcSpan {
                         start,
                         end: value.location().end,
@@ -2240,6 +2278,7 @@ where
                 }))
             } else {
                 Ok(Some(CallArg {
+                    implicit: false,
                     location: value.location(),
                     value,
                     label: None,
@@ -2951,6 +2990,7 @@ fn is_reserved_word(tok: Token) -> bool {
             | Token::Todo
             | Token::Try
             | Token::Type
+            | Token::Use
     ]
 }
 
@@ -2978,6 +3018,7 @@ pub fn make_call(
             ParserArg::Hole { location, label } => {
                 num_holes += 1;
                 CallArg {
+                    implicit: false,
                     label,
                     location,
                     value: UntypedExpr::Var {